requirements.txt
# Data Manipulation & Visualization
pandas==2.2.2
numpy==2.1.0
seaborn==0.13.2
matplotlib==3.9.0
scipy==1.13.0

# Machine Learning & Tracking
scikit-learn==1.5.0
xgboost==2.1.0
mlflow==2.16.0

# Configuration & Environment
pyyaml==6.0.2
python-dotenv==1.0.1

# Interface & LLM
streamlit==1.38.0
openai==1.40.0

# Testing & Development
pytest==8.3.0
ipykernel==6.29.0
